added enemy healer

+++ Assets/_Only Game Assets/Scripts/HealthSystem/HealthSystem.cs
@@ -2,8 +2,8 @@ using UnityEngine;
 
 public class HealthSystem : MonoBehaviour
 {
-    public int maxHealth = 100;
-    public int currentHealth;
+    public float maxHealth = 100;
+    public float currentHealth;
     public HealthUI healthUI;
     void Start()
     {
@@ -15,7 +15,7 @@ public class HealthSystem : MonoBehaviour
         }
     }
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         currentHealth -= amount;
         currentHealth = Mathf.Max(currentHealth, 0); // prevent negative values
@@ -29,7 +29,7 @@ public class HealthSystem : MonoBehaviour
         }
     }
 
-    public void Heal(int amount)
+    public void Heal(float amount)
     {
         currentHealth += amount;
         if (currentHealth > maxHealth)
